Add LOG_BFREE_RELOG to prevent complex log block retire and replay
To retire log block is a bit complex.

For example, when can we free log block? Usually past log records are
releasable on rollup. But, if the record type was bfree after rollup
(say LOG_BFREE_ON_ROLLUP), we can't free that log block until after
rollup was committed. Because bfree of LOG_BFREE_ON_ROLLUP is still
not applied on this rollup (it is flused to bitmap data on next rollup).

So, we don't need to replay almost all logs, but we have to replay
LOG_BFREE_ON_ROLLUP on the past log blocks before latest rollup.

To prevent this complexity, we are logging LOG_BFREE_ON_ROLLUP on this
rollup as LOG_BFREE_RELOG again. With this, we can release
LOG_BFREE_ON_ROLLUP, and furthermore, we don't need to keep past log
blocks before this rollup at all. (log block itself is also logged as
LOG_BFREE_RELOG to free)

To do above, this patch adds LOG_BFREE_RELOG at first.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>