Fix freeblocks for atomic commit
Now, we are updating the freeblocks in sb on delta.  It means
freeblocks and bitmap data (this is written on only rollup) is not
same state.  This makes hard to replay defered bfree.

Because replay has to track whether delta updated freeblocks or not.
For example, we are replaying rollup, the log_bfree means we have to
update freeblocks and bitmap data. But if there is delta after rollup,
freeblocks is already uptodate, so we have to update only bitmap
data for that log_bfree.

To avoid this complexity, this introduces LOG_FREEBLOCKS, and
obsoletes sb->super.freeblocks.  LOG_FREEBLOCKS contains freeblocks
count, and it is logging the freeblocks of bitmap state on rollup.

With this, LOG_FREEBLOCKS is always same with latest rollup state.
And we always replay log_bfree for updating both of bitmap and
freeblocks.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>