Invalidate the buffers which is truncated
We have to invalidate the buffers if truncated. To do it, this adds
truncate_buffers_range(), it makes buffer as empty (i.e. no data
contain. so, bread will read from disk).

We may be able to free buffer actually though, it just leaves buffer
on hash, because buffer->count may not be 0 here. On new_buffer(),
buffers will be reclaimed soon or later.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>