orphan, start orphan management
This just adds orphan logs and orphan table. And almost nothing add
functionality for orphan.

The basic design is following, and in the orphan series, it will be
implemented: (FIXME: sb->otable might not be best, please optimize)

LOG_ORPHAN_ADD and LOG_ORPHAN_DEL are log records of frontend
operation for orphan state. With it, we don't need any write to FS
except log blocks. If the orphan is short life, it will be handled by
this.

However, if the orphan is long life, it can make log blocks too long.
So, to prevent it, if orphan inodes are still living until rollup, we
store those inum into sb->otable. With it, we can obsolete log blocks.

On replay, we can know the inum of orphan inodes yet not destroyed by
checking sb->otable, LOG_ORPHAN_ADD, and LOG_ORPHAN_DEL. (Note, orphan
inum of LOG_ORPHAN_ADD can be destroyed by same inum of LOG_ORPHAN_DEL).

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>