Save delta to buffer->b_state for kernel
We have to know whether buffer is stabilized or not. And delta number
is used for it.

But, kernel buffer_head doesn't have good place to save it. So, this
uses buffer->b_state hack to save delta.

Hack:
 - Map 0-3 to 1-4 for delta number.
 - Kernel buffer helpers can't set "delta" and "dirty bit" atomicity,
   so we use 0 as special number.
 - If we read 0 delta for dirty buffer, it means "delta is not set
   yet". So, we retry until see delta number.

I.e. we do like the following

tux3_mark_buffer_dirty()
{
	mark_buffer_dirty(buffer)
	while (1) {
		old = buffer->b_state;
		new |= delta;
		state = cmpxchg(buffer->b_state, old, new);
		if (old == state)
			break;
	}
}

tux3_get_delta()
{
	while (delta == 0)
		delta = get_delta(buffer);
	return delta;
}

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>